Bug fix for #7 - when we move where past a select the deep copy was getting confused
The AST's were the same as if we did them by hand or not. But when we did it by hand, the object links remained. When it was done by our code, a deep copy (to be safe) si done, and object's, while contents are identical, aren't identical in objects. So .rep's weren't present that are normally.

The bug that this triggered was that rep's were being recalled out of scope. So there are now tests to make sure that rep's are always in scope when a get_rep call is made. And if not, they are zeroed out. This makes for inefficiencies, however. But still, it works now.

- get_rep carefully checks to see if a rep is in scope
- visit no longer rejects visits if a rep is present
- top_scope and regular scope now interoperate more generally
- re-orged where function simplifier tests occur - they weren't in the right spot.

Fixes #7