Update use_exe_func_adl_server.py
Pass up an error when we can't parse the json - and make sure that the text goes along with it.

As we get more expierence with timesouts/etc./ I'm sure this code will have to be further modified. Sort of like it needs some testing perhaps. LOL

This is work on #22